move tests to package `rollbar`
From package `rollbar_tests`.  So we can do unit test on unexported
functions.
