+++ rollbar/provider_test.go
@@ -20,7 +20,7 @@
  * SOFTWARE.
  */
 
-package rollbar_test
+package rollbar
 
 import (
 	"fmt"
@@ -29,7 +29,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 	"github.com/rollbar/terraform-provider-rollbar/client"
-	"github.com/rollbar/terraform-provider-rollbar/rollbar"
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 	"github.com/stretchr/testify/suite"
@@ -67,7 +66,7 @@ type AccSuite struct {
 
 func (s *AccSuite) SetupSuite() {
 	// Setup testing
-	s.provider = rollbar.Provider()
+	s.provider = Provider()
 	s.providers = map[string]*schema.Provider{
 		"rollbar": s.provider,
 	}
