+++ rollbar/resource_project_test.go
@@ -20,7 +20,7 @@
  * SOFTWARE.
  */
 
-package rollbar_test
+package rollbar
 
 import (
 	"fmt"
